Set up mm.h
