+++ buildrules/linker.rules
@@ -11,6 +11,8 @@ dump: $(OBJDIR)/helloos.bin
 	objdump -b binary -D -mi8086 $<
 $(OBJDIR)/boot.img: $(OBJDIR)/helloos.bin
 	dd if=$< of=$@ bs=512 count=1024
+run.x86.4mb: $(OBJDIR)/boot.img
+	qemu-kvm -cpu 486 -m 4 -fda $< -net none -no-kvm
 run.x86.256mb: $(OBJDIR)/boot.img
 	qemu-kvm -cpu 486 -m 256 -fda $< -net none -no-kvm
 run.x86: $(OBJDIR)/boot.img
