+++ hypervisor/mm.c
@@ -1,13 +1,5 @@
-#define MEMMAN_FREES    4090   /*about 32KB*/
+#include "mm.h"
 
-struct FREEINFO {
-    unsigned int addr, size;
-};
-
-struct MEMMAN {
-    int frees, maxfrees, lostsize, losts;
-    struct FREEINFO free[MEMMAN_FREES];
-};
 
 void memman_init(struct MEMMAN *man)
 {
