+++ hypervisor/start.c
@@ -136,7 +136,7 @@ void drawing_desktop()
     if (ret != 0) {
         putfont8_string(vram,xsize, 28, 48, COL8_FFFFFF,font.Bitmap , "Memman Free 2 Failed");
     }
-    sprintf(buf,"MEMORY %d MB. %dKB Free.", memtotal / (1024*1024), memman_total(memman) / 1024);
+    sprintf(buf,"MEMORY %d MB. %dMB Heap Free.", memtotal / (1024*1024), memman_total(memman) / (1024*1024));
     putfont8_string(vram,xsize, 8, 8, COL8_FFFFFF,font.Bitmap , "Hack Week 0x10!!!");
     putfont8_string(vram,xsize, 8, 28, COL8_FFFFFF,font.Bitmap , buf);
